[1/x] Add Initial Blog Mode (#82)
* Add initial setup for blog mode

* Separate nav into its own page

* Add blogNav

* Add navigation back

* Simplify blog nav

* Add setup for blog header

* Add simple naive setup for blog sidebar

* Use generated context for authors

* Use _context

* Add context setup to frontend

* Move blog content into its own file

* Remove _context

* Address all comments
