+++ client/package.json
@@ -51,6 +51,7 @@
     "brotli-size": "^4.0.0",
     "clean-css": "^4.2.3",
     "clsx": "^1.1.1",
+    "date-and-time": "^2.4.1",
     "deasync": "^0.1.20",
     "debounce": "^1.2.0",
     "dlv": "^1.1.3",
