+++ client/src/layouts/NavSidebar/BlogNav.tsx
@@ -1,0 +1,10 @@
+export function BlogNav() {
+  return <>
+    <a className="mt-9 group flex font-semibold text-sm leading-6 text-slate-700 hover:text-slate-900 dark:text-slate-200 dark:hover:text-white" href="/blog">
+      <svg viewBox="0 -9 3 24" className="overflow-visible mr-3 text-slate-400 w-auto h-6 group-hover:text-slate-600 dark:group-hover:text-slate-300">
+        <path d="M3 0L0 3L3 6" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"></path>
+      </svg>
+      See more posts
+    </a>
+  </>
+}
